Add the ability to use edges of rank 2 and update Dockerfile
In homogeneous Markov chains, edge weights do not depend on time.
For them, it is enough to use 2D arrays for edges instead of 3D arrays.
Add the support for them and implement a helper function `get_edge_slice`.

Ubuntu 24.04 LTS has been released.
Its repository contains CLang 18 and other tools we need.
Use it instead of Debian Sid for convenience and stability.